Build a palindrome.js
You have two strings, a and b. Find a string, s, such that:

s can be expressed as s = 8a+ st where 8, is a non-empty substring of a and so is a non-empty substring of b.
$ is a palindromic string.
The length of $ is as long as possible.
For each of the q pairs of strings (a and b) received as input, find and print strings, on a
new line. If you’re able to form more than one valid string $;, print whichever one comes
first alphabetically. If there is no valid answer, print -1 instead.
Input Format
The first line contains a single integer, q, denoting the number of queries. The subsequent
lines describe each query over two lines:
The first line contains a single string denoting a.
The second line contains a single string denoting b.
Output Format
For each pair of strings (a and b), find some s; satisfying the conditions above and print it
on a new line. If there is no such string, print -1 instead.

Sample Input

3
bac
bac
abc
def
jdfh
fds

Sample Output

aba
-1
dfhfd

Explanation
We perform the following three queries:

Concatenate s₁ = “a” with s = “ba” to create s = “aba”.
We’re given a = “abc” and sa = “def”; because both strings are composed of unique characters, we cannot use them to form a palindromic string. Thus, we print -1.
Concatenate s = “dfh” with s¿ = “fd” to create s = “dfhfd”. Note that we chose these particular substrings because the length of string s must be maximal.